pyproject.toml
[project]
name = "api-detailiing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.14",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
]
